Use `if ! defined` instead of ensure_resource()
It might be the case that a project's platform may change (intentionally
or otherwise).  In that case, catalog compilation will fail if we use
`ensure_resource()` on the `sentry::source::project` Exec.

Use `if ! defined` instead, to ensure that once a project has been
created, any changes won't break the catalog.