Allow projects to be created in teams
The create_project.py script had previously accepted a `-t` flag to
specify the team to which the project should be assigned. This commit
allows create_project.py to create that team if it does not exist.

Note that the `-o` flag allows the specification of an organization, but
the script **does not** create new organizations at this time.

The `sentry::source::export` and `sentry::source::project` defined types
have been updated to pass organization and team.

For projects created outside of the default organization+team combo, the
DSN cache file will include the organization and team in the directory
name. These values are `uriescape()`d to ensure that spaces and other
things are supported.