Use resource containment, ftw
By using private classes, we can ensure that the parameters passed into
them will be those set for the top-level `sentry` class itself.  This,
in turn, means we don't need to use resource-like class invocaitons.  We
just need to ensure that resource ordering is properly defined, and that
earlier classes notify later classes.